+++ src/hooks/index.ts
@@ -4,4 +4,6 @@ export * from './useInitVisynApp';
 export * from './useSyncedRef';
 export * from './useVisynUser';
 export * from './useSetRef';
-export * from './useDeep';
+export * from './useDeepComparison';
+export * from './useDeepEffect';
+export * from './useDeepMemo';
